+++ R/f_s3_rendoboots.R
@@ -63,8 +63,13 @@ confint.rendo.boots <- function(object, parm, level=0.95, ...){
   # bootstrap replications is not much larger than 1/min(alpha,1-alpha) because there are not enough
   # theta* values from which to extract the desired quantile
   # (e.g., one cannot plausibly estimate a 99 % quantile given fewer than 100 values)."
-  if(ncol(object$boots.params) < 1/(1-level))
-    stop("Not enough bootstraps (<1/(1-level)) were performed to derive the confidence interval!", call. = FALSE)
+
+  if(any(rowSums(!is.na(object$boots.params)) < 1/(1-level)))
+    stop("Not enough bootstraps (<1/(1-level)) were performed to derive the confidence interval or too many estimates were NA!", call. = FALSE)
+
+  # warn if any bootstrap estimates are NA
+  if(anyNA(object$boots.params))
+    warning("Confidence intervals might be unreliable because there are NAs in the bootstrapped parameter estimates.", call. = FALSE)
 
   # This largely follows stats:::confint.lm to exhibit the exact same behavior