+++ NEWS.md
@@ -1,3 +1,17 @@
+# CHANGES IN REndo 2.2.1
+
+## SIGNIFICANT USER-VISIBLE CHANGES
+* To tweak the lmer model fit in `multilevelIV` to their linking, users can supply a parameter `lmer.control`
+
+## NEW FEATURES
+* Warning in `confint` if there are NAs in the bootstrapped estimates for `copulaCorrection`
+* Updated vignette
+
+## BUG FIXES
+* The coefficient estimates for `multilevelIV` are more consistent and independent of data sorting due to different standard settings for fitting `lmer`
+* Bootstrapped parameter estimates for `copualCorrection` result in fewer NAs when using L-BFGS-B as optimization method
+
+
 # CHANGES IN REndo 2.2.0
 
 ## SIGNIFICANT USER-VISIBLE CHANGES
@@ -10,11 +24,11 @@
 * New data was generated for `dataMultilevelIV`
 
 ## BUG FIXES
-* The sigma matrix in `latentIV` is constructed as in the paper by Ebbes what improves results. Special thanks to Jordan Lawson for investigating and pointing this out!
-* In the `latentIV`, the parameter for group membership (`theta5`) is  transformed back and now reported correctly.
+* The sigma matrix in `latentIV` is constructed as in the paper by Ebbes et al. what improves results. Special thanks to Jordan Henderson for investigating and pointing this out!
+* In `latentIV`, the parameter for group membership (`theta5`) is now transformed back and reported correctly.
 * The vcov matrix for `latentIV` is corrected for the transformation in `theta5`.
 * The bootstrapping in `copulaCorrection` case 1 now creates samples of the same length as the original data
-* The (percentile) confidence intervals and vcov for results from `copulaCorrection` now are derived with bootstrapping
+* The (percentile) confidence intervals and vcov for results from `copulaCorrection` now rely on bootstraping
 
 
 # CHANGES IN REndo 2.1.0